fdasfad

+++ AlgorithmsFFLib/src/MS2Ion.h
@@ -19,7 +19,7 @@ class ALGORITHMSFFLIB_EXPORTS MS2Ion {
 
 public:
 
-    double mz = -1.0;
+    float mz = -1.0;
     float intensity = -1.0;
     IRT iRT = -1.0;
     QString ionLabel;
